+++ src/trx/game/output/lights/tr12.c
@@ -5,6 +5,7 @@
 #include <trx/game/const.h>
 #include <trx/game/output.h>
 #include <trx/game/output/lights.h>
+#include <trx/game/output/lights/fog_bulbs.h>
 #include <trx/game/output/lights/priv.h>
 #include <trx/gl/utils.h>
 #include <trx/version.h>
@@ -313,5 +314,6 @@ const LIGHTING_MODEL g_LightingModelTR12 = {
     .add_dynamic_light_rgb = M_AddDynamicLightRGB,
     .upload_cpu_light = M_UploadCPULight,
     .upload_own_light = M_UploadOwnLight,
+    .prepare_scene = Output_FogBulbs_PrepareScene,
     .shader_variant = 0,
 };